Trabajo Terminado

+++ Script.sql
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS Producto (
     id_producto INT AUTO_INCREMENT PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     descripcion VARCHAR(255),
-    precio DECIMAL(10,2) NOT NULL,
+    precio DOUBLE NOT NULL,
     stock INT NOT NULL CHECK (stock >= 0)
 );
 
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS Cliente (
     nombre VARCHAR(100) NOT NULL,
     email VARCHAR(100) NOT NULL,
     telefono VARCHAR(20) NOT NULL,
-    dinero int NOT NULL,
+    dinero DOUBLE NOT NULL,
     CONSTRAINT chk_dinero_positivo CHECK (dinero >= 0)
 );
 
@@ -47,7 +47,7 @@ CREATE TABLE IF NOT EXISTS DetalleVenta (
     id_venta INT NOT NULL,
     id_producto INT NOT NULL,
     cantidad INT NOT NULL CHECK (cantidad > 0),
-    precio_unitario DECIMAL(10,2) NOT NULL,
+    precio_unitario DOUBLE NOT NULL,
     FOREIGN KEY (id_venta) REFERENCES Venta(id_venta) ON DELETE CASCADE,
     FOREIGN KEY (id_producto) REFERENCES Producto(id_producto) ON DELETE CASCADE
 );
@@ -58,7 +58,7 @@ CREATE PROCEDURE RegistrarDetalleVenta(
     IN p_idVenta INT,       
     IN p_idProducto INT,      
     IN p_cantidad INT,        
-    IN p_precioUnitario DECIMAL(10, 2), 
+    IN p_precioUnitario DOUBLE, 
     OUT p_estado INT          
 )
 BEGIN
@@ -89,10 +89,10 @@ DELIMITER ;
 Delimiter //
 
 CREATE FUNCTION CalcularTotalVenta(p_id_venta INT)
-RETURNS DECIMAL(10,2)
+RETURNS DOUBLE
 DETERMINISTIC
 BEGIN
-    DECLARE total DECIMAL(10,2);
+    DECLARE total DOUBLE;
 
     SELECT SUM(cantidad * precio_unitario) INTO total 
     FROM detalleVenta
@@ -102,4 +102,3 @@ BEGIN
 END //
 
 DELIMITER ;
-
